+++ shared/output/gapic/templates/showcase/test/google/showcase/v1beta1/messaging_paths_test.rb
@@ -38,11 +38,17 @@ class ::Google::Showcase::V1beta1::Messaging::ClientPathsTest < Minitest::Test
         config.credentials = grpc_channel
       end
 
+      path = client.blurb_path user_id: "value0", legacy_user_id: "value1", blurb_id: "value2"
+      assert_equal "users/value0/profile/blurbs/legacy/value1~value2", path
+
+      path = client.blurb_path user_id: "value0", blurb_id: "value1"
+      assert_equal "users/value0/profile/blurbs/value1", path
+
       path = client.blurb_path room_id: "value0", blurb_id: "value1"
       assert_equal "rooms/value0/blurbs/value1", path
 
-      path = client.blurb_path user_id: "value0", blurb_id: "value1"
-      assert_equal "user/value0/profile/blurbs/value1", path
+      path = client.blurb_path room_id: "value0", legacy_room_id: "value1", blurb_id: "value2"
+      assert_equal "rooms/value0/blurbs/legacy/value1.value2", path
     end
   end